+++ chatroom/chatdialog.cc
@@ -1,0 +1,13 @@
+#include "chatdialog.hpp"
+
+#include "ui_chatdialog.h"
+
+ChatDialog::ChatDialog(QWidget* parent) : QDialog(parent), ui(new Ui::ChatDialog)
+{
+  ui->setupUi(this);
+}
+
+ChatDialog::~ChatDialog()
+{
+  delete ui;
+}