+++ chatroom/mainwindow.hpp
@@ -26,6 +26,7 @@ QT_END_NAMESPACE
 class LoginDialog;
 class RegisterDialog;
 class ResetPasswordDialog;
+class ChatDialog;
 class MainWindow : public QMainWindow
 {
   Q_OBJECT
@@ -38,6 +39,7 @@ public slots:
   void SlotSwitchRegister();
   void SlotSwitchLogin();
   void SlotSwitchResetPassword();
+  void SlotSwitchChat();
 
 private:
   Ui::MainWindow* ui;
@@ -47,5 +49,6 @@ private:
   LoginDialog* _login_dialog;
   RegisterDialog* _register_dialog;
   ResetPasswordDialog* _reset_password_dialog;
+  ChatDialog* _chat_dialog;
 };
 #endif  // MAINWINDOW_HPP